requirements.txt
# Core dependencies
streamlit>=1.28.0
python-docx>=0.8.11
pdfplumber>=0.10.3

# Embeddings and vector store
sentence-transformers>=2.2.2
chromadb>=0.4.18
torch>=2.0.0

# Text processing
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken>=0.5.0

# HTTP client for Ollama
requests>=2.31.0
httpx>=0.25.0

# Image processing
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# For table extraction
pandas>=2.0.0
openpyxl>=3.1.0
